Create README.md
This is a Simple User Registration Script in PHP and MySQL. By using this user can easily register, filling username, email, password and date of birth fields. After submitting the form, the data will be stored in a database table. So that next time user can login with the login credentials which are used while user registration.